+++ baybe/parameters/categorical.py
@@ -13,13 +13,7 @@ from baybe.parameters.base import _DiscreteLabelLikeParameter
 from baybe.parameters.enum import CategoricalEncoding
 from baybe.parameters.validation import validate_unique_values
 from baybe.settings import active_settings
-from baybe.utils.conversion import nonstring_to_tuple
-
-
-def _convert_values(value, self, field) -> tuple[str, ...]:
-    """Sort and convert values for categorical parameters."""
-    value = nonstring_to_tuple(value, self, field)
-    return tuple(sorted(value, key=lambda x: (str(type(x)), x)))
+from baybe.utils.conversion import normalize_convertible2str_sequence
 
 
 def _validate_label_min_len(self, attr, value) -> None:
@@ -38,8 +32,10 @@ class CategoricalParameter(_DiscreteLabelLikeParameter):
     # object variables
     _values: tuple[str | bool, ...] = field(
         alias="values",
-        converter=Converter(_convert_values, takes_self=True, takes_field=True),  # type: ignore
-        validator=(
+        converter=Converter(  # type: ignore[misc,call-overload]  # mypy: Converter
+            normalize_convertible2str_sequence, takes_self=True, takes_field=True
+        ),
+        validator=(  # type: ignore[arg-type]  # mypy: validator tuple
             validate_unique_values,
             deep_iterable(
                 member_validator=(instance_of((str, bool)), _validate_label_min_len),