+++ baybe/symmetries/__init__.py
@@ -1,0 +1,13 @@
+"""Symmetry classes for expressing invariances of the modeling process."""
+
+from baybe.symmetries.base import Symmetry
+from baybe.symmetries.dependency import DependencySymmetry
+from baybe.symmetries.mirror import MirrorSymmetry
+from baybe.symmetries.permutation import PermutationSymmetry
+
+__all__ = [
+    "DependencySymmetry",
+    "MirrorSymmetry",
+    "PermutationSymmetry",
+    "Symmetry",
+]